% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSGI.R
\name{avg.dist.calc}
\alias{avg.dist.calc}
\title{to calculate distance between gradated factors}
\usage{
avg.dist.calc(info, r_squared_thresh = 0.6, minimum.fctr = 10)
}
\arguments{
\item{info}{LSGI (local.traj.preprocessing) output}

\item{r_squared_thresh}{R-squared threshold: only plot gradient with higher R-squared than this value}

\item{minimum.fctr}{if some factors have too few gradients, ignore it}
}
\value{
returns a heatmap of distance between gradated factors
}
\description{
to calculate distance between gradated factors
}
