% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSGI.R
\name{calc.local.linear}
\alias{calc.local.linear}
\title{calculates local gradients at each grid point}
\usage{
calc.local.linear(
  grids,
  dist.to.grid,
  latent.embeddings,
  spatial_coords,
  n.cells.per.meta
)
}
\arguments{
\item{grids}{output of get.grid.coords}

\item{dist.to.grid}{distance between each cell/spot to each grid point}

\item{latent.embeddings}{embeddings used to calculate gradient with}

\item{spatial_coords}{a dataframe with cell/spot as row and axis (X, Y) as column}

\item{n.cells.per.meta}{the size of local group; it is suggested to use 20-50 as a start}
}
\value{
returns local linear gradient information
}
\description{
calculates local gradients at each grid point
}
