% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSGI.R
\name{get.grid.coords}
\alias{get.grid.coords}
\title{this function is called by others to compute the localizations of grid points}
\usage{
get.grid.coords(spatial_coords, n.grids.scale = 5)
}
\arguments{
\item{spatial_coords}{a dataframe with cell/spot as row and axis (X, Y) as column}

\item{n.grids.scale}{number of grids to use equals to (total number of cells)/n.grids.scales}
}
\value{
returns a dataframe with grid points as rows and coordinates as columns
}
\description{
this function is called by others to compute the localizations of grid points
}
