% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSGI.R
\name{get.ind.rsqrs}
\alias{get.ind.rsqrs}
\title{to gather grid level information from the LSGI result file.}
\usage{
get.ind.rsqrs(info)
}
\arguments{
\item{info}{LSGI output}
}
\value{
returns a dataframe with all gradient (for every factor) at all locations (for every grid point)
}
\description{
to gather grid level information from the LSGI result file.
}
