% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSGI.R
\name{local.traj.preprocessing}
\alias{local.traj.preprocessing}
\title{main function of the LSGI package}
\usage{
local.traj.preprocessing(
  spatial_coords,
  embeddings,
  n.grids.scale = 5,
  n.cells.per.meta = 25
)
}
\arguments{
\item{spatial_coords}{a dataframe with cell/spot as row and axis (X, Y) as column}

\item{embeddings}{embeddings used to calculate gradient with}

\item{n.grids.scale}{number of grids to use equals to (total number of cells)/n.grids.scales}

\item{n.cells.per.meta}{the size of local group; it is suggested to use 20-50 as a start}
}
\value{
returns local linear gradient information and other basic information of intermediate steps of this analysis
}
\description{
main function of the LSGI package
}
