% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSGI.R
\name{optimize.arrow}
\alias{optimize.arrow}
\title{to optimize arrow length for visualization}
\usage{
optimize.arrow(grid.info, scale.factor = 50)
}
\arguments{
\item{grid.info}{part of the output of local.traj.preprocessing}

\item{scale.factor}{scale factor for lengths of arrows; 50 is good for visium data as tested}
}
\value{
returns updated grid.info
}
\description{
to optimize arrow length for visualization
}
