% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSGI.R
\name{plt.dist.heat}
\alias{plt.dist.heat}
\title{to plot heatmap of distance between gradated factors}
\usage{
plt.dist.heat(dist.mat)
}
\arguments{
\item{dist.mat}{output of avg.dist.calc}
}
\value{
returns a heatmap of distance between gradated factors
}
\description{
to plot heatmap of distance between gradated factors
}
