% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSGI.R
\name{plt.factor.gradient.ind}
\alias{plt.factor.gradient.ind}
\title{to plot individual factor superimposed with embeddings/levels}
\usage{
plt.factor.gradient.ind(info, fctr, r_squared_thresh, arrow.length.scale = 1)
}
\arguments{
\item{info}{LSGI (local.traj.preprocessing) output}

\item{fctr}{factor to plot}

\item{r_squared_thresh}{R-squared threshold: only plot gradient with higher R-squared than this value}

\item{arrow.length.scale}{for adjusting arrow length in a flexible way}
}
\value{
returns a plot with individual factor superimposed with embeddings/levels
}
\description{
to plot individual factor superimposed with embeddings/levels
}
