% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSGI.R
\name{plt.factors.gradient.ind}
\alias{plt.factors.gradient.ind}
\title{to plot multiple factors together on the spatial map}
\usage{
plt.factors.gradient.ind(
  info,
  r_squared_thresh,
  sel.factors = NULL,
  minimum.fctr = 5,
  arrow.length.scale = 1
)
}
\arguments{
\item{info}{LSGI (local.traj.preprocessing) output}

\item{r_squared_thresh}{R-squared threshold: only plot gradient with higher R-squared than this value}

\item{sel.factors}{factor to plot; by default all}

\item{minimum.fctr}{if some factors have too few gradients, ignore it}

\item{arrow.length.scale}{for adjusting arrow length in a flexible way}
}
\value{
returns a plot with multiple selected factors on the spatial map
}
\description{
to plot multiple factors together on the spatial map
}
